Backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic
PyPDF2
python-multipart
openai